+++ OpenAI-DotNet/Realtime/CreateResponseRequest.cs
@@ -1,6 +1,5 @@
 ﻿// Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System;
 using System.Text.Json.Serialization;
 
 namespace OpenAI.Realtime
